// gravitee-apim-vitest/model/apis.ts
export interface Api {
  id: string;
  contextPath: string;
  description: string;
  endpoint: string;
  name: string;
  version: string;
  state: ApiState;
  visibility: ApiVisibility;
  lifecycle_state: ApiLifecycleState;
  created_at: number;
  updated_at: number;
  owner: string;
  workflow_state: string;
  labels?: string[];
}

export interface Error {
  message: string;
}
// ...
export enum ApiVisibility {
  PUBLIC = 'PUBLIC',
  PRIVATE = 'PRIVATE',
}

export enum ApiLifecycleState {
  CREATED = 'CREATED',
  PUBLISHED = 'PUBLISHED',
  UNPUBLISHED = 'UNPUBLISHED',
  DEPRECATED = 'DEPRECATED',
  ARCHIVED = 'ARCHIVED',
}

export enum ApiWorkflowState {
  IN_REVIEW = 'IN_REVIEW',
  REVIEW_OK = 'REVIEW_OK',
}

export enum ApiState {
  STOPPED = 'STOPPED',
  STARTED = 'STARTED',
}